requirements.txt
streamlit==1.24.1
pypdf2==3.0.1
langchain==0.0.228
python-dotenv==1.0.0
faiss-cpu==1.7.4
openai==0.27.8
tiktoken==0.4.0
